Extract out timestamp
