+++ journal-server/Cargo.toml
@@ -9,3 +9,4 @@ edition = "2018"
 [dependencies]
 journal = { version = "0.1.0", path = "../journal" }
 tokio = { version = "1.5.0", features = ["full"] }
+chrono = {version = "0.4.19"}
