+++ journal/Cargo.toml
@@ -7,7 +7,6 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-chrono = {version = "0.4.19"}
 libc = {version = "0.2.77"}
 nix = { version = "0.18.0" }
 journalctl-sys = { version = "0.1.0", path = "../journalctl-sys" }
